Improve connector hatch waila info

+++ src/main/java/tectech/thing/metaTileEntity/hatch/MTEHatchDataOutput.java
@@ -167,7 +167,7 @@ public class MTEHatchDataOutput extends MTEHatchDataConnector<QuantumDataPacket>
 
         NBTTagCompound tag = accessor.getNBTData();
         currenttip.add(
-            translate("tt.keyphrase.Computation_Transmitting", GTUtility.formatNumbers(tag.getInteger("computation"))));
+            translate("tt.keyphrase.Computation_Transmitting", GTUtility.formatNumbers(tag.getLong("computation"))));
     }
 
     @Override
